I think my doc is tested to a good extent
Hi, If you feel like adding this link on your ReadMe which is displayed by default when someone clicks on your repo. I did not give this pull request before as I wanted my docs to be tested more. Now as Erwan and a few others including me have gone through this doc a few times, I think this doc is ready.